Debugged confusion plotting

+++ Utility/include/plotutil.h
@@ -17,6 +17,7 @@
 #include <TLatex.h>
 
 #include <ChargedAnalysis/Utility/include/vectorutil.h>
+#include <ChargedAnalysis/Utility/include/rootutil.h>
 
 namespace PUtil{
     void SetStyle();
@@ -28,7 +29,7 @@ namespace PUtil{
     void DrawLegend(TPad* pad, TLegend* legend, const int& nColumns);
     void DrawShapes(TCanvas* canvas, TH1* bkg, TH1* sig);
     
-    void DrawConfusion(const std::vector<int>& trueLabel, const std::vector<int>& predLabel, const std::vector<std::string>& classNames, const std::string& outDir);
+    void DrawConfusion(const std::vector<long>& trueLabel, const std::vector<long>& predLabel, const std::vector<std::string>& classNames, const std::string& outDir);
 
     std::string GetChannelTitle(const std::string& channel);
     std::string GetLumiTitle(const std::string& lumi);
